% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_theme.R
\name{change_theme}
\alias{change_theme}
\title{Changes the RStudio-theme}
\usage{
change_theme(theme = NULL)
}
\arguments{
\item{theme}{A specific theme-name (e.g. Pastel on dark, Textmate (default), etc.).}
}
\description{
If theme=NULL, then the function switches between the "Pastel on dark"
and "Textmate (default)" themes.
}
\details{
Changes the RStudio-theme in two ways. Either by giving a specific RStudio-theme or
just changing between dark and light mode,if theme=NULL (the default).
}
\author{
Dennis Freuer
}
