% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat}
\alias{cormat}
\title{Heatmap and ordered table of pairwise correlations}
\usage{
cormat(df, print_cors = FALSE, rnd = NULL)
}
\arguments{
\item{df}{A data frame (categorical variables are allowed and will be ignored).}

\item{print_cors}{If TRUE, the pairwise correlation will be printed to the console.}

\item{rnd}{The number of digits of the decimal place.}
}
\value{
Returns a list of length two - plot and data frame ordered by correlations.
}
\description{
This function is an easy way to get ordered bivariate correlations in a decreasing way and
a visualisation in the form of a heatmap. It is very useful for detection of the strongest and weakest
correlations, especially if the given dataset contains many variables. The function divides
automatically the data into numerical and categorical variables and applies the pearson
correlation on the numerical variables (categorical variables are allowed and will be ignored).
}
\author{
Dennis Freuer
}
