% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_path.R
\name{current_path}
\alias{current_path}
\title{Returns the path to folder of the current R-script}
\usage{
current_path()
}
\value{
Returns the current path as a string.
}
\description{
Returns the current path to the folder, where the current R-script is stored.
}
\author{
Dennis Freuer
}
