% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_cov}
\alias{df_cov}
\title{Covariate values.}
\format{A tibble with 12 rows and 4 variables:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{VARIABLE}{Variable}
  \item{VALUE}{Value of the variable}
}}
\usage{
df_cov
}
\description{
A dataset containing covariate information.
}
\keyword{datasets}
