% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_cov_evid4}
\alias{df_cov_evid4}
\title{A covariate data set to be used with EVID=4.}
\format{A data frame with 119 rows and 6 variables:
\describe{
  \item{ID}{Individual}
  \item{PERIOD}{Period}
  \item{TIMEPOINT}{Timepoint, in hours}
  \item{TIME}{Time, in hours}
  \item{VARIABLE}{Variable}
  \item{VALUE}{Value of the variable}
}}
\usage{
df_cov_evid4
}
\description{
A dataset containing covariate information.
}
\keyword{datasets}
