% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_cov_start}
\alias{df_cov_start}
\title{Starting covariate data set.}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{ID}{Individual}
  \item{VARIABLE}{Variable}
  \item{VALUE}{Value of the variable}
}}
\usage{
df_cov_start
}
\description{
A dataset containing covariate information.
}
\keyword{datasets}
