% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_cov_time_dependent_start}
\alias{df_cov_time_dependent_start}
\title{A covariate data set to be used with time dependent covariates.}
\format{A data frame with 6 rows and 4 variables:
\describe{
  \item{ID}{Individual}
  \item{VARIABLE}{Variable}
  \item{VALUE}{Value of the variable}
  \item{TIME}{Time, in hours}
}}
\usage{
df_cov_time_dependent_start
}
\description{
A dataset containing time dependent covariates.
}
\keyword{datasets}
