% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dose}
\alias{df_dose}
\title{A dose data set.}
\format{A data frame with 12 rows and 3 variables:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{AMT}{Dose, in mg}
}}
\usage{
df_dose
}
\description{
A dataset containing dose information.
}
\keyword{datasets}
