% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dose_datetime}
\alias{df_dose_datetime}
\title{A dose data set including datetimes.}
\format{A data frame with 5 rows and 12 variables:
\describe{
  \item{ID}{Individual}
  \item{TRT}{Treatment label}
  \item{DOSE}{Dose, in mg}
  \item{PERIOD}{Period}
  \item{DAY}{Day of adminsitration}
  \item{AMT}{Dose, in mg}
  \item{DATETIME}{Dta ein datetime format}
  \item{TIMEPOINT}{Timepoint}
  \item{COHORT}{Cohort}
  \item{FORM}{Drug form}
  \item{TREATMENT}{Treatment}
  \item{FOOD}{Food status}
}}
\usage{
df_dose_datetime
}
\description{
A dataset containing dose information in datetime format.
}
\keyword{datasets}
