% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dose_evid4}
\alias{df_dose_evid4}
\title{A dose data set to be used with EVID=4.}
\format{A data frame with 418 rows and 10 variables:
\describe{
  \item{ID}{Individual}
  \item{PERIOD}{Period}
  \item{TIMEPOINT}{Timepoint}
  \item{TIME}{Time, in hours}
  \item{AMT}{Dose, in mg}
  \item{TRT}{Treatment label}
  \item{DAY}{Day of adminsitration}
  \item{SEQUENCE}{Sequence}
  \item{TRT2}{Treatment}
  \item{EVID}{Evid value}
}}
\usage{
df_dose_evid4
}
\description{
A dataset containing dosing information.
}
\keyword{datasets}
