% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dose_optional_columns}
\alias{df_dose_optional_columns}
\title{A dose data set to be used with optional columns.}
\format{A data frame with 4 rows and 6 variables:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{AMT}{Dose, in mg}
  \item{OCC}{Occasion}
  \item{TIMEPOINT}{Timepoint}
  \item{TRT}{Treatment}
}}
\usage{
df_dose_optional_columns
}
\description{
A dataset containing dosing information.
}
\keyword{datasets}
