% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_dose_start}
\alias{df_dose_start}
\title{A dose data set example.}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{AMT}{Dose, in mg}
}}
\usage{
df_dose_start
}
\description{
A dataset containing dosing information.
}
\keyword{datasets}
