% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_extra_times}
\alias{df_extra_times}
\title{An extra times data set example.}
\format{A data frame with 251 rows and 1 variable:
\describe{
  \item{TIME}{Time, in hours}
}}
\usage{
df_extra_times
}
\description{
A dataset containing extra times.
}
\keyword{datasets}
