% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_extra_times_datetime}
\alias{df_extra_times_datetime}
\title{An extra times data set example with datetime format.}
\format{A data frame with 20 rows and 1 variable:
\describe{
  \item{ID}{Individual}
  \item{DATETIME}{Datetime}
  \item{TIMEPOINT}{Timepoint}
}}
\usage{
df_extra_times_datetime
}
\description{
A dataset containing extra times in datetime format.
}
\keyword{datasets}
