% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_extra_times_parent_evid4}
\alias{df_extra_times_parent_evid4}
\title{An extra times parent data set to be used with EVID=4.}
\format{A data frame with 770 rows and 3 variable:
\describe{
  \item{PERIOD}{Period}
  \item{TIMEPOINT}{Timepoint}
  \item{TIME}{Time, in hours}
}}
\usage{
df_extra_times_parent_evid4
}
\description{
A dataset containing extra times for an hypothetical parent drug.
}
\keyword{datasets}
