% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_extra_times_time}
\alias{df_extra_times_time}
\title{An extra times data set example.}
\format{A data frame with 1040 rows and 3 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{TIMEPOINT}{Timepoint}
}}
\usage{
df_extra_times_time
}
\description{
A dataset containing extra times.
}
\keyword{datasets}
