% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_parent_evid4}
\alias{df_parent_evid4}
\title{A pharmacokinetic parent data set to be used with EVID=4.}
\format{A data frame with 1359 rows and 7 variables:
\describe{
  \item{ID}{Individual}
  \item{PERIOD}{Period}
  \item{TIMEPOINT}{Timepoint}
  \item{TIME}{Time, in hours}
  \item{DV}{Drug concentration, in mg/L}
  \item{TIMEDAY}{Timeday}
  \item{DAY}{Day of adminsitration}
}}
\usage{
df_parent_evid4
}
\description{
A dataset containing pharmacokinetic information for an hypothetical parent drug.
}
\keyword{datasets}
