% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pd_start}
\alias{df_pd_start}
\title{An starting pharmacoynamic data set example.}
\format{A tibble with 6 rows and 3 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{DV}{Response, ng/mL}
}}
\usage{
df_pd_start
}
\description{
A dataset containing pharmacodynamic observations.
}
\keyword{datasets}
