% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pk}
\alias{df_pk}
\title{A pharmacokinetic data set.}
\format{A tibble with 132 rows and 4 variable:
\describe{
  \item{ID}{Individual}
  \item{TIMEPOINT}{Timepoint}
  \item{TIME}{Time, in hours}
  \item{DV}{Drug concentration, ng/mL}
}}
\usage{
df_pk
}
\description{
A dataset containing pharmacokinetic information.
}
\keyword{datasets}
