% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pk_datetime}
\alias{df_pk_datetime}
\title{A pharmacokinetic data set example in datetime format.}
\format{A data frame with 65 rows and 7 variable:
\describe{
  \item{ID}{Individual}
  \item{DV}{Response, ng/mL}
  \item{DATETIME}{Datetime}
  \item{TIMEPOINT}{Timepoint}
  \item{DAY}{Day}
  \item{PERIOD}{Period}
  \item{BLQ}{I a BLQ?}
  \item{LLOQ}{Lower limit of quantification, ng/mL}
}}
\usage{
df_pk_datetime
}
\description{
A dataset containing pharmacokinetic information.
}
\keyword{datasets}
