% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pk_metabolite}
\alias{df_pk_metabolite}
\title{A pharmacokinetic data set of metabolite data.}
\format{A data frame with 10 rows and 3 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{DV}{Drug concentration, ng/mL}
}}
\usage{
df_pk_metabolite
}
\description{
A dataset containing pharmacokinetic information for an hypothetical metabolite.
}
\keyword{datasets}
