% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pk_optional_columns}
\alias{df_pk_optional_columns}
\title{A pharmacokinetic data set to be used with optional columns.}
\format{A data frame with 12 rows and 5 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{DV}{Drug concentration, ng/mL}
  \item{OCC}{Occasion}
  \item{TIMEPOINT}{Timepoint}
}}
\usage{
df_pk_optional_columns
}
\description{
A dataset containing pharmacokinetic information.
}
\keyword{datasets}
