% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pk_parent}
\alias{df_pk_parent}
\title{A pharmacokinetic data set for an hypothetical parent drug.}
\format{A data frame with 12 rows and 3 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{DV}{Drug concentration, ng/mL}
}}
\usage{
df_pk_parent
}
\description{
A dataset containing pharmacokinetic information.
}
\keyword{datasets}
