% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pk_start}
\alias{df_pk_start}
\title{A pharmacokinetic data set example.}
\format{A tibble with 12 rows and 3 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{DV}{Response, ng/mL}
}}
\usage{
df_pk_start

df_pk_start
}
\description{
A dataset containing pharmacokinetic information.

A dataset containing pharmacokinetic information.
}
\keyword{datasets}
