% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_response1}
\alias{df_response1}
\title{A pharmacodynamic data set.}
\format{A data frame with 6 rows and 3 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{DV}{Response, ng/mL}
}}
\usage{
df_response1
}
\description{
A dataset containing pharmacodynamic information for response 1.
}
\keyword{datasets}
