% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_response3}
\alias{df_response3}
\title{A pharmacodynamic data set.}
\format{A data frame with 6 rows and 3 variable:
\describe{
  \item{ID}{Individual}
  \item{TIME}{Time, in hours}
  \item{DV}{Response, in hours}
}}
\usage{
df_response3
}
\description{
A dataset containing pharmacodynamic information for response 3.
}
\keyword{datasets}
