% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcol.R
\name{ggcol}
\alias{ggcol}
\title{Creates ggplot2 default colours}
\usage{
ggcol(n = 3, h = c(0, 360) + 15)
}
\arguments{
\item{n}{An integer value representing the desired number of colours.}

\item{h}{A numeric vector for adjusting the colours.}
}
\value{
Returns a character vector of length n that contains the colour code.
}
\description{
Creates the colour code for the first n ggplot2 default colours.
}
\author{
Dennis Freuer
}
