% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfluct.R
\name{ggfluct}
\alias{ggfluct}
\title{Fluctuation plot}
\usage{
ggfluct(tbl, type = "size", boxFill = "grey90", floor = 0,
  ceiling = max(table$freq, na.rm = TRUE), xlab = "", ylab = "")
}
\arguments{
\item{tbl}{A 2-dim table (cross-table).}

\item{type}{Represents the type of boxes within the plot.}

\item{boxFill}{String that represents the colour of boxes within the plot.}

\item{floor}{Controls the min-size of the boxes within the plot.}

\item{ceiling}{Controls the max-size of the boxes within the plot.}

\item{xlab}{Character string representing the label of  the x-axis.}

\item{ylab}{Character string representing the label of  the y-axis.}
}
\description{
This function generates a fluctuation plot in ggplot-style.
}
\author{
Dennis Freuer
}
