% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic.R
\name{ggmosaic}
\alias{ggmosaic}
\title{Mosaic plot in ggplot-style}
\usage{
ggmosaic(data, x, y, useNA = "no", ...)
}
\arguments{
\item{data}{A dataset.}

\item{x}{.}

\item{y}{.}

\item{useNA}{If "no", the default.}

\item{...}{Further inputs.}
}
\description{
Creates a nice-looking mosaic plot in ggplot-style.
}
\author{
Dennis Freuer
}
