% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{google}
\alias{google}
\title{Direct Google-search}
\usage{
google(what)
}
\arguments{
\item{what}{Character string for searching within Google search engine.}
}
\description{
This function allows a Google-search directly from R-shell.
}
\author{
Dennis Freuer
}
