% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkgs.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Loads and installs multiple packages at once}
\usage{
load_pkgs(pkgs, install_new_pgs = TRUE)
}
\arguments{
\item{pkgs}{vector of package names as character strings.}

\item{install_new_pgs}{Should unavailable packages be installed
automatically using the install.packages() function? Default is TRUE.}
}
\description{
Loads and installs (if necessary) several packages at once.
}
\author{
Dennis Freuer
}
