% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkd_lm_results.R
\name{mkd_lm_results}
\alias{mkd_lm_results}
\title{Creates a markdown document showing the linear regression results}
\usage{
mkd_lm_results(
  lm_list,
  dat = NULL,
  mkd_path = NULL,
  open = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{lm_list}{A (named) list object including linear regression.}

\item{mkd_path}{Path to the folder, where the markdown document should be stored.
If NULL, the current path will be used.}

\item{open}{Should the HTML output be opened immediately after saving? Default is FALSE.}

\item{overwrite}{Should an already existing file be overwritten? Default is TRUE.}
}
\value{
Save a HTML document without returning anything.
}
\description{
Creates a markdown document for a (named) list of linear models (from rms::ols() or lm())
showing clearly the summary statistics and estimates in tables that can be presented.
}
\details{
This function is also used within the verif_lm() function.
}
\author{
Dennis Freuer
}
