% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkd_lm_verif.R
\name{mkd_lm_verif}
\alias{mkd_lm_verif}
\title{Creates a markdown document to validate linear model assumptions}
\usage{
mkd_lm_verif(verif_lm_list, mkd_path = NULL, open = FALSE, overwrite = TRUE)
}
\arguments{
\item{verif_lm_list}{A list object including either the result or a (named) list 
of multiple results from the verif_lm() function.}

\item{mkd_path}{Path to the folder, where the markdown document should be stored.
If NULL, the current path will be used.}

\item{open}{Should the HTML output be opened immediately after saving? Default is FALSE.}

\item{overwrite}{Should an already existing file be overwritten? Default is TRUE.}
}
\value{
Save a HTML document without returning anything.
}
\description{
Creates a markdown document for a (named) list of linear model validations from verif_lm() function output.
}
\details{
This function is also used within the verif_lm() function.
}
\author{
Dennis Freuer
}
