% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Plots several ggplots in one}
\usage{
multiplot(plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{List of ggplots.}

\item{cols}{Number of columns within the desired plot.}

\item{layout}{User defined layout.}
}
\description{
This function plots several ggplots in one single graphic with user defined number of columns.
}
\author{
Dennis Freuer
}
