% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_all_files_in.R
\name{open_all_files_in}
\alias{open_all_files_in}
\title{Opens all disred files in a folder}
\usage{
open_all_files_in(path_to_folder, file_type)
}
\arguments{
\item{path_to_folder}{The path to the folder as character string.}

\item{file_type}{The file type that should be opened.It is also possible to 
use a specific string indicating the file names of interest.}
}
\description{
Opens all files with a specific file type in a folder.
}
\author{
Dennis Freuer
}
