% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendir.R
\name{opendir}
\alias{opendir}
\title{File access direclty from R}
\usage{
opendir(dir = getwd())
}
\arguments{
\item{dir}{Path, that should be opened with the current working directory as default.}
}
\description{
Short function for opening either a file directory within any operating system (Windows, MAC OS, etc.) 
or a website (only working with the complete url) from R directly.
}
\author{
Dennis Freuer
}
