% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missings.R
\name{plot_missings}
\alias{plot_missings}
\title{Positions and number of missings within a dataset}
\usage{
plot_missings(d, print_res = TRUE, decreasing = TRUE)
}
\arguments{
\item{d}{Data frame.}

\item{print_res}{Controls whether the column names and counted missings should be printed at the console.}

\item{decreasing}{If TRUE, the variable names will be shown at the console
in a decreasing way. In the same order as in the data frame otherwise.}
}
\value{
Returns a list of length two - plot and data frame ordered by correlations.
}
\description{
This function plots the position of missings within a given dataset and
print the variables with the number of missings in an ordered way.
}
\author{
Dennis Freuer
}
