% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_na.R
\name{rowwise_na}
\alias{rowwise_na}
\title{Rowwise missing values}
\usage{
rowwise_na(dt)
}
\arguments{
\item{dt}{data frame or data table with missing values in any row.}
}
\value{
Returns a numerical vector of lenghth nrow(d).
}
\description{
This function counts for each row of a dataset the number of missing values (NAs).
}
\author{
Dennis Freuer
}
