% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnames.R
\name{varnames}
\alias{varnames}
\title{Variable names of a dataset with a substring of interest}
\usage{
varnames(d, what, case_sensitive = FALSE)
}
\arguments{
\item{d}{Dataset that contains the variable names of interest.}

\item{what}{Character vector of substrings are assumed to be in the variable names.}

\item{case_sensitive}{If FALSE, the default, R doesn't distinquish between upper and lower case.}
}
\value{
Returns a character vector with variable names of interest.
}
\description{
This function searches within a given dataset for variable names that contain one or more specific substrings.
}
\author{
Dennis Freuer
}
