% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_lm.R
\name{verif_lm}
\alias{verif_lm}
\title{Advanced residual plots and tests for linear models}
\usage{
verif_lm(mdl, results = c("print", "return"))
}
\arguments{
\item{mdl}{A linear model of class lm.}

\item{results}{Choose "print", if results should be printed, but not returned.
Choose "return" if results should be returned in a list, but not printed.}
}
\value{
Returns, if desired, a list of length 6 with an summary of tests,
all calculated tests and plots.
}
\description{
Function for creating residual plots in ggplot style and all necessary tests,
so that the assumptions of linear regression models can be verified.
}
\author{
Dennis Freuer
}
