% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmosaic.R
\name{ggmosaic}
\alias{ggmosaic}
\title{Mosaic plot in ggplot-style}
\usage{
ggmosaic(data, xvar, yvar, useNA = "no", ...)
}
\arguments{
\item{data}{A dataset.}

\item{xvar}{name of the first variable}

\item{yvar}{name of the second variable}

\item{useNA}{Should NAs be considered (regarding to the table() function)}

\item{...}{Further inputs.}
}
\description{
Creates a nice-looking mosaic plot in ggplot-style.
}
\author{
Dennis Freuer
}
