% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle.R
\name{puzzle}
\alias{puzzle}
\title{Assembling Pharmacometric Databases}
\usage{
puzzle(directory = NULL, order, pk = list(name = NULL, data = NULL),
  dose = list(name = NULL, data = NULL), cov = list(name = NULL, data =
  NULL), pd = list(name = NULL, data = NULL), extratimes = list(name =
  NULL, data = NULL), nm = list(name = NULL), coercion = list(name =
  NULL, sep = ","), optionalcolumns = NULL, fillcolumns = NULL,
  nocoercioncolumns = NULL, norepeatcolumns = NULL, initialindex = 0,
  na.strings = "N/A", arrange = "ID,TIME,CMT,desc(EVID)",
  datetimeformat = "\%Y-\%m-\%d \%H:\%M:\%S", timeunits = "hours",
  timezone = Sys.timezone(), ignore = "C", missingvalues = ".",
  parallel = T, verbose = F)
}
\arguments{
\item{directory}{path to your directory}

\item{order}{define the absorption order, can be 0, 1, c(0,1), or c(1,1)}

\item{pk}{define the required file containing the pk information. It can be a .csv or an .xlsx file}

\item{dose}{define the required file containing the dose information. It can be a .csv, an .xlsx file or an R object.}

\item{cov}{define the optional file containing the covariate information. It can be a .csv, an .xlsx file or an R object.}

\item{pd}{define the optional file containing the pd information. It can be a .csv, or a .xlsx file.}

\item{extratimes}{define the optional file containing the additional times. It can be a .csv, or a .xlsx file.}

\item{nm}{name of output file generated by puzzle}

\item{optionalcolumns}{define optional columns}

\item{initialindex}{define the lower category of categorical covariates}

\item{arrange}{define how the columns should be arranged}

\item{timeunits}{define time units if needed}

\item{parallel}{define parallel zero + first order absorption}

\item{dateformat}{define date format if needed}

\item{timeformat}{define time format if needed}
}
\value{
a .csv file
}
\description{
Build a dataset in the specific format required for the NONMEM analysis from basic tabulated files
}
\examples{

puzzle(directory=file.path(inDirectory,"Theophylline"),order=1,
             pk=list(name="pk.csv"), dose=list(name="dose.csv"), cov=list(name="cov.csv"), extratimes=list(name="addltimes.csv"),
             nm=list(name="nm1.csv"),
             optionalcolumns="TIMEPOINT")

nm=puzzle(directory=file.path(inDirectory,"Theophylline"),order=1,
             pk=list(name="pk.csv"), dose=list(name="dose.csv"), cov=list(name="cov.csv"), extratimes=list(name="addltimes.csv"),
             optionalcolumns="TIMEPOINT")
write.csv(nm, file=file.path(inDirectory,"Theophylline","nm1b.csv"),row.names=F,quote=F,na=".")

puzzle(directory=file.path(inDirectory,"inAssembling/0. Poster"),order=c(1,1),
             pk=list(name="pk.csv"), pd=list(name="pd.csv"), dose=list(name="dose.csv"), cov=list(name="cov.csv"), extratimes=list(name="addltimes.csv"),
             coercion=list(name="nm11.txt",sep=";"),
             nm=list(name="nm11.csv"),
             optionalcolumns="OCC")

puzzle(directory=file.path(inDirectory,"inAssembling/1. Basic"), order=0,
             pk=list(name="pk.csv"), pd=list(name="pd.csv"), dose=list(name="dose.csv"), cov=list(name="cov.csv"),
             nm=list(name="nm_ZeroOrder.csv"))

puzzle(directory=file.path(inDirectory,"inAssembling/1. Basic"), order=1,
             pk=list(name="pk.csv"), pd=list(name="pd.csv"), dose=list(name="dose.csv"), cov=list(name="cov.csv"),
             nm=list(name="nm_FirstOrder.csv"))

puzzle(directory=file.path(inDirectory,"inAssembling/1. Basic"), order=c(0,1),
             pk=list(name="pk.csv"), pd=list(name="pd.csv"), dose=list(name="dose.csv"), cov=list(name="cov.csv"),
             nm=list(name="nm_ParallelZeroAndFirstOrder.csv"))

puzzle(directory=file.path(inDirectory,"inAssembling/2. Multiple analytes and effects"), order=1,
             pk=list(name="pk2.xlsx"), pd=list(name="pd3.xlsx"), dose=list(name="dose.csv"), cov=list(name="cov.csv"),
             nm=list(name="nm_MultipleAnalytesandEffects.csv"))

puzzle(directory=file.path(inDirectory,"inAssembling/3. Optional columns"),order=1,
             pk=list(name="pk_OptionalColumns.csv"), dose=list(name="dose_OptionalColumns.csv"), cov=list(name="cov.csv"),
             coercion=list(name="nm_OptionalColumns.txt"),
             nm=list(name="nm_OptionalColumns.csv"),
             optionalcolumns=c("OCC","TIMEPOINT","TRT"), fillcolumns="TRT")

puzzle(directory=file.path(inDirectory,"inAssembling/4. EVID4"),order=1,
             pk=list(name="pk.xlsx"), dose=list(name="dose.csv"), cov=list(name="cov.csv"),
             extratimes=list(name="addltimes.xlsx"),
             nm=list(name="nm_1_full.csv"),
             coercion=list(name="nm_1_full.txt"),
             optionalcolumns=c("PERIOD","TIMEPOINT","TRT"), fillcolumns="TRT", arrange="ID,PERIOD,TIME,desc(MDV),CMT")

puzzle(directory=file.path(inDirectory,"inAssembling/5. extratimes mixte"),order=1,
             pk=list(name="pk_2017-11-23.csv"), dose=list(name="doses_2017-11-23.csv"),
             nm=list(name="nm.csv"),
             optionalcolumns=c("TIMEPOINT"), datetimeformat="\%Y-\%m-\%d \%H:\%M")

puzzle(directory=file.path(inDirectory,"inAssembling/5. extratimes mixte"),order=1,
             pk=list(name="pk_2017-11-23.csv"), dose=list(name="doses_2017-11-23.csv"), extratimes=list(name="extratimes_2017-11-23_TIME.csv"),
             nm=list(name="nmt.csv"),
             optionalcolumns=c("TIMEPOINT"), datetimeformat="\%Y-\%m-\%d \%H:\%M")

puzzle(directory=file.path(inDirectory,"inAssembling/5. extratimes mixte"),order=1,
             pk=list(name="pk_2017-11-23.csv"), dose=list(name="doses_2017-11-23.csv"), extratimes=list(name="extratimes_2017-11-23_DATETIME.csv"),
             nm=list(name="nmdt.csv"),
             optionalcolumns=c("TIMEPOINT"), datetimeformat="\%Y-\%m-\%d \%H:\%M")

puzzle(directory=file.path(inDirectory,"inAssembling/5. extratimes mixte"),order=1,
             pk=list(name="pk_2017-11-23.csv"), dose=list(name="doses_2017-11-23.csv"), extratimes=list(name="extratimes_2017-11-23.csv"),
             nm=list(name="nmtdt.csv"),
             optionalcolumns=c("TIMEPOINT"), datetimeformat="\%Y-\%m-\%d \%H:\%M")
}
